[Feature-WIP] Fix bugs from previous commit
Identified these bugs when testing out the changes in the demo
Laravel JSON API application.

At the same time, simplified the paginator classes so they are
directly injected with the encoding parameters. This is now
possible because the encoding parameters for the current request
are registered directly in the service container.