[Feature] Add Eloquent hydrator
Add an Eloquent hydrator which handles the majority of use cases
for hydrating Eloquent models.

As part of this, update the Eloquent controller so that it invokes
secondary hydration and saving operations if its injected
hydrator implements the `HydratesRelatedInterface`. This allows
the Eloquent hydrator to hydrate belongs-to-many relationships
after the primary model has been created.