[Feature-WIP] Remove search and add pagination validator
Search has been removed and merged into the Eloquent adapter.

A pagination validator has been added so that pagination parameters
can now be validated using a Laravel validator. This allows for
the max range of the size of a page to be set in the validator,
for example by using 'integer|between:1,50` as the page size
rule.