[Bugfix] Nest page meta by default
This matches the previous version of this package, that always
nested the page meta. By adding this back in, the upgrade path
is better. The improvement here is nesting can be turned off if
required.

Nesting the meta is a better default anyway because it avoids
collisions with other meta keys that are being included in the
document's top-level meta key.