[Refactor] Update composer dependencies
No longer requiring the entire Laravel framework and instead
only requiring the Illuminate packages required.

The `illuminate/routing` package has been made a dev dependency
because it is not used in Lumen. It is required to use the
controllers and routing classes within this package, but they
would only be used in a Laravel appication anyway.