[Feature] Improve resource query parameter validation
When requesting a resource as the primary data, the `sort`,
`filter` and `page` parameters are now not allowed by default,
because they do not apply in these scenarios.

In addition, when creating a query parameter validator for these
scenarios, any rules that relate to `sort`, `filter` and `page`
parameters are filtered out because those rules do not need to
be applied. This means that the `required` rule can be used for
any of these paramters and the rule will only be hit in scenarios
where sort, filter and page can exist (searching resources).

Fixes #128