[Refactor] Sort out the argument order for controller hooks
The record is now passed as the first argument except for the
creating hook (as the record does not exist for that hook).
This makes a lot more sense as in almost all cases the record
is likely to be used (e.g. for dispatching events or jobs) rather
than the JSON API resource object.