[Feature] Allow resources to not have controllers
A resource can now be flagged as not having a controller when
register resource routes. This is done by setting the
`controller` option to `false`. This is allowed because the
generic JSON API controller can now handle any resource, so
it is only necessary to have a specific controller if you need
to either implement hooks or override the default behaviour.