[Refactor] Make controllers optional by default
As the `JsonApiController` is able to handle any resource without
customisation, the controller is not actually required. This
commit makes the controller optional by default, as this helps
set up a resource quicker and means a developer only needs to
generate a controller if they actually need it.

The old behaviour of using a controller with the same name as the
resource can be achieved by passing `true` as the controller
option when registering resource routes.