[Refactor] Move related implementation to abstract resource adapter
This was previously on the Eloquent adapter but is actually not
unique to Eloquent so can be moved into the parent abstract class.
This means there is one less method to implement on a non-Eloquent
adapter and means custom adapters can use the same pattern as the
Eloquent adapter for defining relationships.