[Feature-WIP] Add failing tests for extra route parameters
If there are route parameters before the JSON API route parameters,
the current implementation results in errors because the
JSON API controller is not expecting additional route parameters
to be passed into its action methods. This commit adds a failing
test for this scenario using a wildcard sub-domain as the scenario.