[Feature] Do not require inverse relationship adapter
If there are no encoding parameters that need to be processed
by an inverse adapter for a to-many relationship, we can return
the Eloquent relationship directly from the primary record.
This allows us to have relationships to resources for which
only a schema exists, e.g. if the resource only ever appears in
a compound document or resource relationship endpoint.