[Refactor] Deprecate Eloquent schemas
Eloquent schemas are not really required, as they add a lot of
extra processing that is not necessary as generic schemas are
very simple to write. They are therefore deprecated as generic
schemas will be a lot more efficient.

This commit modifies the stub for the Eloquent schema so that
it is a generic schema instead. But to help, the id method is
already completed and the created-at and updated-at attributes
are included. The developer can modify this as needed.