[Refactor] Append type to class name when using not-by-resource
The resolver now expects the type to be appended to the classname
when using not-by-resource resolution. I.e.
App\JsonApi\Adapters\PostAdapter is now expected instead of
App\JsonApi\Adapters\Post.

The old behaviour can be maintained by setting the by-resource
option to a string: `false-0.x`.

This matches Laravel behaviour better, as per the discussion in
Issue #176.