[Refactor] Use Eloquent route key as default resource id
Changes the default resource id to use `Model::getRouteKey`/
`Model::getRouteKeyName` instead of `Model::getKey`/
`Model::getKeyName`.

This is because the JSON API key is used in the route, so the
route key is more appropriate as the default. This is easily
overridden in the schema and adapter if needed.

See discussion in #183.

Closes #184