[Feature] Allow filtering of resource read requests
This feature allows a request for a specific resource to
be filtered. E.g.:

```
GET /api/posts/1?filter['published']=1
```

Returns post 1 if it is published.

The same is also now possible on to-one relations:

```
GET /api/comments/1/post?filter['published']=1
```

Returns the post that is associated to comment 1 if that post is
published.

Closes #218