Remove resource provider
It is not really possible to implement a default resource adapter/
validator logic because it is too application specific. Therefore
this needs to be removed and installing apps need to generate
their own resource classes. Only a schema is provided to extend.

Migrations are now only run if the application opts in to them.