[Bugfix] Reject request with client id when not supported
The JSON API spec states that a server must reject a request to
create a resource with a client-generated ID, if the resource
does not support client-generated IDs. The reject must be a
`403 Forbidden` response.

This change fixes this by defining on the validators class whether
the resource supports client-generated IDs.

Closes #259