[Feature] Add query parameters to test helpers
Adds a query parameters argument to all the JSON API test
request helpers. This makes it easier to test JSON API requests
that use query parameters, as they will now be converted from
an API into a HTTP query string and appended to the URI being
tests.

This change also reduces the depedency on the API's settings for
the tests, by constructing the URL from a base URL set on the
test case - rather than via the API object. This makes the tests
better as they become independent from the API's settings (i.e.
prevents false-positives by the test not relying on the thing it
is testing).