[Feature] Allow non-camel case adapter relation method names.
Although developers should follow the PSR1 standard of camel
casing method names, there are some developers who follow
the Eloquent attribute standard of snake casing attributes
when defining relationships.

This commit allows an adapter to have its relation method name
in exactly the same format as the JSON API field name. I.e.
a field name of `user_history` and a relation method called
`user_history`.

Otherwise it falls back to the PSR1 standard and expects the
method name to be the camel case version of the field name.

See #315 for discussion.