[Feature] Add Eloquent filtering using scopes
Adds a `filterWithScopes()` method to the Eloquent adapter,
that can be used to map JSON API filters to Eloquent model
local scopes.

This is opt-in, as in the developer must call `filterWithScopes()`
within their `filter()` method to use it. It has been added
to the generated Eloquent adapter as a suggested default
implementation.