[Tests] Fix broken tests
Two tests were failing because of the use of posts and videos
resources for the comment commentable relationship - and each
model has a different primary key column (id vs uuid). This test
setup was a bit unrealistic as morph-to relations are more likely
to have the same id type on the different models.